+++ src/components/_classes/component/editForm/Component.edit.validation.js
@@ -86,20 +86,20 @@ export default [
             <pre>valid = (input === 'Joe') ? true : 'Your name must be "Joe"';</pre>
           </small>`
       },
-      {
-        type: 'well',
-        components: [
-          {
-            weight: 100,
-            type: 'checkbox',
-            label: 'Secret Validation',
-            tooltip: 'Check this if you wish to perform the validation ONLY on the server side. This keeps your validation logic private and secret.',
-            description: 'Check this if you wish to perform the validation ONLY on the server side. This keeps your validation logic private and secret.',
-            key: 'validate.customPrivate',
-            input: true
-          }
-        ]
-      }
+      // {
+      //   type: 'well',
+      //   components: [
+      //     {
+      //       weight: 100,
+      //       type: 'checkbox',
+      //       label: 'Secret Validation',
+      //       tooltip: 'Check this if you wish to perform the validation ONLY on the server side. This keeps your validation logic private and secret.',
+      //       description: 'Check this if you wish to perform the validation ONLY on the server side. This keeps your validation logic private and secret.',
+      //       key: 'validate.customPrivate',
+      //       input: true
+      //     }
+      //   ]
+      // }
     ]
   },
   {
@@ -156,48 +156,83 @@ export default [
       {
         type: 'htmlelement',
         tag: 'div',
-        content: `
-          <p>This allows you to set different custom error messages for different errors
-          (in contrast to “Custom Error Message”, which only allows you to set one
-          error message for all errors). E.g.</p>
-
+        content: `<p>This allows you to set different custom error messages for different errors (in contrast to “Custom Error Message”, which only allows you to set one error message for all errors). E.g.</p>
 <pre>{
-  "required": "{<span/>{ field }} is required. Try again.",
-  "maxLength": "{<span/>{ field }} is too long. Try again."
+  "required": "{<span>{ field }</span>} is required. Try again.",
+  "maxLength": "{<span>{ field }</span>} is too long. Try again."
 }</pre>
+<p>You can set the following keys (among others):</p>
+<ul>
+<li><span>r</span>equired</li>
+<li><span>m</span>in</li>
+<li><span>m</span>ax</li>
+<li><span>m</span>inLength</li>
+<li><span>m</span>axLength</li>
+<li><span>m</span>inWords</li>
+<li><span>m</span>axWords</li>
+<li><span>i</span>nvalid_email</li>
+<li><span>i</span>nvalid_date</li>
+<li><span>i</span>nvalid_day</li>
+<li><span>i</span>nvalid_regex</li>
+<li><span>m</span>ask</li>
+<li><span>p</span>attern</li>
+<li><span>c</span>ustom</li>
+</ul>
+<p>Depending on the error message some of the following template variables can be used in the script:</p>
+<ul>
+<li><code>{<span>{ <span>f</span>ield }</span>}</code> is replaced with the label of the field.</li>
+<li><code>{<span>{ <span>m</span>in }</span>}</code></li>
+<li><code>{<span>{ <span>m</span>ax }</span>}</code></li>
+<li><code>{<span>{ <span>l</span>ength }</span>}</code></li>
+<li><code>{<span>{ <span>p</span>attern }</span>}</code></li>
+<li><code>{<span>{ <span>m</span>inDate }</span>}</code></li>
+<li><code>{<span>{ <span>m</span>axDate }</span>}</code></li>
+<li><code>{<span>{ <span>m</span>inYear }</span>}</code></li>
+<li><code>{<span>{ <span>m</span>axYear }</span>}</code></li>
+<li><code>{<span>{ <span>r</span>egex }</span>}</code></li>
+</ul>`
+//         content: `
+//           <p>This allows you to set different custom error messages for different errors
+//           (in contrast to “Custom Error Message”, which only allows you to set one
+//           error message for all errors). E.g.</p>
 
-          <p>You can set the following keys (among others):</p>
-          <ul>
-            <li>r<span/>equired</li>
-            <li>m<span/>in</li>
-            <li>m<span/>ax</li>
-            <li>m<span/>inLength</li>
-            <li>m<span/>axLength</li>
-            <li>m<span/>inWords</li>
-            <li>m<span/>axWords</li>
-            <li>i<span/>nvalid_email</li>
-            <li>i<span/>nvalid_date</li>
-            <li>i<span/>nvalid_day</li>
-            <li>i<span/>nvalid_regex</li>
-            <li>m<span/>ask</li>
-            <li>p<span/>attern</li>
-            <li>c<span/>ustom</li>
-          </ul>
+// <pre>{
+//   "required": "{<span/>{ field }} is required. Try again.",
+//   "maxLength": "{<span/>{ field }} is too long. Try again."
+// }</pre>
 
-          <p>Depending on the error message some of the following template variables can be used in the script:</p>
-          <ul>
-           <li><code>{<span/>{ f<span/>ield }}</code> is replaced with the label of the field.</li>
-           <li><code>{<span/>{ m<span/>in }}</code></li>
-           <li><code>{<span/>{ m<span/>ax }}</code></li>
-           <li><code>{<span/>{ l<span/>ength }}</code></li>
-           <li><code>{<span/>{ p<span/>attern }}</code></li>
-           <li><code>{<span/>{ m<span/>inDate }}</code></li>
-           <li><code>{<span/>{ m<span/>axDate }}</code></li>
-           <li><code>{<span/>{ m<span/>inYear }}</code></li>
-           <li><code>{<span/>{ m<span/>axYear }}</code></li>
-           <li><code>{<span/>{ r<span/>egex }}</code></li>
-          </ul>
-        `
+//           <p>You can set the following keys (among others):</p>
+//           <ul>
+//             <li>r<span/>equired</li>
+//             <li>m<span/>in</li>
+//             <li>m<span/>ax</li>
+//             <li>m<span/>inLength</li>
+//             <li>m<span/>axLength</li>
+//             <li>m<span/>inWords</li>
+//             <li>m<span/>axWords</li>
+//             <li>i<span/>nvalid_email</li>
+//             <li>i<span/>nvalid_date</li>
+//             <li>i<span/>nvalid_day</li>
+//             <li>i<span/>nvalid_regex</li>
+//             <li>m<span/>ask</li>
+//             <li>p<span/>attern</li>
+//             <li>c<span/>ustom</li>
+//           </ul>
+
+//           <p>Depending on the error message some of the following template variables can be used in the script:</p>
+//           <ul>
+//            <li><code>{<span/>{ f<span/>ield }}</code> is replaced with the label of the field.</li>
+//            <li><code>{<span/>{ m<span/>in }}</code></li>
+//            <li><code>{<span/>{ m<span/>ax }}</code></li>
+//            <li><code>{<span/>{ l<span/>ength }}</code></li>
+//            <li><code>{<span/>{ p<span/>attern }}</code></li>
+//            <li><code>{<span/>{ m<span/>inDate }}</code></li>
+//            <li><code>{<span/>{ m<span/>axDate }}</code></li>
+//            <li><code>{<span/>{ m<span/>inYear }}</code></li>
+//            <li><code>{<span/>{ m<span/>axYear }}</code></li>
+//            <li><code>{<span/>{ r<span/>egex }}</code></li>
+//           </ul>
+//         `
       }
     ]
   }