+++ zip_dist.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+zip -r formio.dist.4.14.10.zip dist